So there's some odd behaviour in my tech combo...
Immer w/ Next.js's SSR w/ this React version was causing intermittant
attempts to access frozen objects.

`produce()` is returning readonly types--likely due to a change in
Immer v.7 that defaults autoFreeze on all the time, even in prod.

This is a little sad, as deep-freezing in dev is great.